+++ app/components/TextAnimation.tsx
@@ -1,228 +1,71 @@
-import React, { useEffect, useRef } from 'react';
-import * as THREE from "three";
+import React, { useRef } from 'react';
+import { Canvas, useFrame } from '@react-three/fiber';
+import { Text, Float, Sparkles } from '@react-three/drei';
+import * as THREE from 'three';
 
-import { OrbitControls } from "three/examples/jsm/controls/OrbitControls.js";
-import getBgSphere from "./getBgSphere";
-
-import { TTFLoader } from "three/examples/jsm/loaders/TTFLoader.js";
-import { Font } from "three/examples/jsm/loaders/FontLoader.js";
-import { TextGeometry } from "three/examples/jsm/geometries/TextGeometry.js";
-
-import { LineMaterial } from "three/examples/jsm/lines/LineMaterial.js";
-import { Line2 } from "three/examples/jsm/lines/Line2.js";
-import { LineGeometry } from "three/examples/jsm/lines/LineGeometry.js";
-
-interface TextAnimationProps {
-  message?: string;
-  fontPath?: string;
-  bgHue?: number;
+interface AnimatedTextProps {
+  text: string;
+  position?: [number, number, number];
+  color?: string;
+  fontSize?: number;
 }
 
-const TextAnimation: React.FC<TextAnimationProps> = ({
-  message = "Three.js",
-  fontPath = "./fonts/ChakraPetch-Bold.ttf",
-  bgHue = 0.6
+const AnimatedText: React.FC<AnimatedTextProps> = ({
+  text,
+  position = [0, 0, 0],
+  color = 'white',
+  fontSize = 1,
 }) => {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const sceneRef = useRef<{
-    scene: THREE.Scene;
-    camera: THREE.PerspectiveCamera;
-    renderer: THREE.WebGLRenderer;
-    controls: OrbitControls;
-    text: THREE.Group | null;
-    animationId: number | null;
-  }>({ 
-    scene: new THREE.Scene(),
-    camera: new THREE.PerspectiveCamera(),
-    renderer: new THREE.WebGLRenderer(),
-    controls: null as unknown as OrbitControls,
-    text: null,
-    animationId: null
+  const meshRef = useRef<THREE.Mesh | null>(null);
+
+  // continuous slow rotation
+  useFrame(() => {
+    if (meshRef.current) {
+      meshRef.current.rotation.y += 0.01;
+      meshRef.current.rotation.x += 0.005;
+    }
   });
 
-  function createOutlines({ font, message }: { font: Font; message: string }) {
-    const strokeGroup = new THREE.Group();
-  
-    let totalDist = 1.0;
-    const lineMaterial = new LineMaterial({
-      color: 0xffffff,
-      linewidth: 3,
-      dashed: true,
-      dashSize: totalDist * 2,
-      gapSize: totalDist * 2,
-      dashOffset: Math.random() * totalDist,
-    });
-  
-    function getStrokeMesh({ shape, i = 0.0 }: { shape: THREE.Shape; i?: number }) {
-      let points = shape.getPoints();
-      let points3d: number[] = [];
-      points.forEach((p) => {
-        points3d.push(p.x, p.y, 0);
-      });
-      const lineGeo = new LineGeometry();
-      lineGeo.setPositions(points3d);
-    
-      totalDist = shape.getLength();
-      lineMaterial.dashSize = totalDist * 2;
-      lineMaterial.gapSize = totalDist * 2;
-      lineMaterial.dashOffset = Math.random() * totalDist;
-      
-      const strokeMesh = new Line2(lineGeo, lineMaterial);
-      strokeMesh.computeLineDistances();
-      let offset = i * 0;
-      strokeMesh.userData.update = (t: number) => {
-        strokeMesh.material.dashOffset = t * (totalDist * 0.1) + offset;
-      };
-      return strokeMesh;
-    }
-  
-    const shapes = font.generateShapes(message, 1);
-    shapes.forEach((s: THREE.Shape, i: number) => {
-      strokeGroup.add(getStrokeMesh({ shape: s, i }));
-  
-      if (s.holes?.length > 0) {
-        s.holes.forEach((h) => {
-          strokeGroup.add(getStrokeMesh({ shape: new THREE.Shape(h.getPoints()), i }));
-        });
-      }
-    });
-    strokeGroup.userData.update = (t: number) => {
-      strokeGroup.children.forEach((c) => {
-        c.userData.update?.(t);
-      });
-    };
-    return strokeGroup;
-  }
-  
-  function createText({ font, message }: { font: Font; message: string }) {
-    const textGroup = new THREE.Group();
-    const props = {
-      font,
-      size: 1,
-      depth: 0.1,
-      curveSegments: 6,
-      bevelEnabled: true,
-      bevelThickness: 0.08,
-      bevelSize: 0.01,
-      bevelOffset: 0,
-      bevelSegments: 2,
-    };
-    const textGeo = new TextGeometry(message, props);
-    textGeo.computeBoundingBox();
-    const centerOffset = -0.5 * (textGeo.boundingBox!.max.x - textGeo.boundingBox!.min.x);
-    const glassMat = new THREE.MeshPhysicalMaterial({
-      roughness: 0.5,
-      transmission: 1.0,
-      transparent: true,
-      thickness: 1.0,
-    });
-    const textMesh = new THREE.Mesh(textGeo, glassMat);
-    textMesh.position.x = centerOffset;
-    textGroup.add(textMesh);
-  
-    const outlineText = createOutlines({ font, message });
-    outlineText.position.set(centerOffset, 0, 0.2);
-    textGroup.add(outlineText);
-  
-    textGroup.userData.update = (t: number) => {
-      let timeStep = t * 0.005;
-      outlineText.userData.update(timeStep);
-    };
-    return textGroup;
-  }
-
-  function initScene(res: any) {
-    const { scene, camera, renderer, controls } = sceneRef.current;
-    const font = new Font(res);
-    const text = createText({ font, message });
-    scene.add(text);
-    sceneRef.current.text = text;
-    animate(0);
-  }
-  
-  function animate(timeStep: number) {
-    const { scene, camera, renderer, text } = sceneRef.current;
-    sceneRef.current.animationId = requestAnimationFrame(animate);
-    if (text) {
-      text.userData.update(timeStep);
-    }
-    renderer.render(scene, camera);
-    sceneRef.current.controls.update();
-  }
-  
-  function loadFont() {
-    const loader = new TTFLoader();
-    loader.load(fontPath, (res: any) => {
-      initScene(res);
-    });
-  }
-
-  function handleWindowResize() {
-    const { camera, renderer } = sceneRef.current;
-    if (containerRef.current) {
-      const width = containerRef.current.clientWidth;
-      const height = containerRef.current.clientHeight;
-      camera.aspect = width / height;
-      camera.updateProjectionMatrix();
-      renderer.setSize(width, height);
-    }
-  }
-
-  useEffect(() => {
-    if (!containerRef.current) return;
-
-    // Setup
-    const width = containerRef.current.clientWidth;
-    const height = containerRef.current.clientHeight;
-    
-    const scene = new THREE.Scene();
-    const camera = new THREE.PerspectiveCamera(75, width / height, 0.1, 1000);
-    camera.position.z = 5;
-    
-    const renderer = new THREE.WebGLRenderer({ antialias: true });
-    renderer.setSize(width, height);
-    containerRef.current.appendChild(renderer.domElement);
-    
-    const controls = new OrbitControls(camera, renderer.domElement);
-    controls.enableDamping = true;
-    
-    const bgSphere = getBgSphere({ hue: bgHue });
-    scene.add(bgSphere);
-    
-    // Store references
-    sceneRef.current = {
-      scene,
-      camera,
-      renderer,
-      controls,
-      text: null,
-      animationId: null
-    };
-    
-    // Initialize
-    loadFont();
-    
-    // Event listeners
-    window.addEventListener('resize', handleWindowResize);
-    
-    // Cleanup
-    return () => {
-      if (sceneRef.current.animationId) {
-        cancelAnimationFrame(sceneRef.current.animationId);
-      }
-      if (containerRef.current && containerRef.current.contains(renderer.domElement)) {
-        containerRef.current.removeChild(renderer.domElement);
-      }
-      window.removeEventListener('resize', handleWindowResize);
-    };
-  }, [bgHue, fontPath, message]);
-
   return (
-    <div 
-      ref={containerRef} 
-      style={{ width: '100%', height: '100%', minHeight: '400px' }}
-    />
+    <Float
+      speed={1.2}           // float speed
+      rotationIntensity={0.8}
+      floatIntensity={1.5}
+    >
+      {/* group both Text and Sparkles together */}
+      <group position={position}>
+        <Text
+          ref={meshRef}
+          color={color}
+          fontSize={fontSize}
+          anchorX="center"
+          anchorY="middle"
+        >
+          {text}
+        </Text>
+
+        <Sparkles
+          count={120}               // number of sparkle particles
+          scale={[fontSize * 2, fontSize * 2, fontSize * 2]}
+          size={1.6 * fontSize}    // sparkle size relative to text
+          speed={0.3}
+          color={color}
+        />
+      </group>
+    </Float>
   );
 };
+
+interface SceneProps {
+  text: string;
+}
+
+const TextAnimation: React.FC<SceneProps> = ({ text }) => (
+  <Canvas camera={{ position: [0, 0, 5], fov: 75 }}>
+    <ambientLight intensity={0.5} />
+    <pointLight position={[10, 10, 10]} />
+    <AnimatedText text={text} />
+  </Canvas>
+);
 
 export default TextAnimation;
