+++ app/components/getBgSphere.js
@@ -1,13 +1,32 @@
-import * as THREE from 'three';
+import * as THREE from "three"
 
-function getBgSphere({ hue = 0.6 }) {
-  const bgGeometry = new THREE.SphereGeometry(50, 32, 32);
-  const bgMaterial = new THREE.MeshBasicMaterial({
-    color: new THREE.Color().setHSL(hue, 0.8, 0.1),
-    side: THREE.BackSide,
-  });
-  const bgSphere = new THREE.Mesh(bgGeometry, bgMaterial);
-  return bgSphere;
+export default function getBgSphere({ hue = 0.565, lightnessMult = 0.015 } = {}) {
+    const bgSphereGeo = new THREE.IcosahedronGeometry(4, 3);
+    const bgSphereMat = new THREE.MeshBasicMaterial({
+        side: THREE.BackSide,
+        vertexColors: true,
+        fog: false
+    });
+    // create an array of colors per vertex
+    const bgSphereColors = [];
+    const len = bgSphereGeo.attributes.position.count;
+    for (let i = 0; i < len; i++) {
+        const z = -bgSphereGeo.attributes.position.getZ(i);
+        const x = bgSphereGeo.attributes.position.getX(i);
+        const y = bgSphereGeo.attributes.position.getY(i);
+        
+        // Add some noise based on position 
+        const noiseValue = Math.sin(x * 5) * Math.cos(y * 5) * 0.1;
+        
+        const { r, g, b } = new THREE.Color().setHSL(
+            (hue + noiseValue) % 1.0,
+            1,
+            z * lightnessMult + 0.2
+        );
+        bgSphereColors.push(r, g, b);
+    }
+    bgSphereGeo.setAttribute('color', new THREE.Float32BufferAttribute(bgSphereColors, 3));
+    const bgSphere = new THREE.Mesh(bgSphereGeo, bgSphereMat);
+
+    return bgSphere;
 }
-
-export default getBgSphere;
