+++ package.json
@@ -21,7 +21,7 @@
     "isbot": "^4.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "three": "^0.146.0",
+    "three": "^0.176.0",
     "three-stdlib": "^2.36.0"
   },
   "devDependencies": {
